refactor(web): replace DevicesManager with api/storage/devices
